requirements.txt
numpy==1.26.4
pandas==2.2.0
yfinance==0.2.36
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
transformers==4.37.2
datasets==2.17.1
sqlalchemy==1.4.50
fredapi==0.5.1
torch==2.2.0
